addClaimant: delete AddClaimantForm, update AddClaimantPage to include form code
